finally got it!

+++ proceed.php
@@ -94,8 +94,8 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         header("Location: manage.php");
     } elseif ($_REQUEST["type"] == "change-depart") {
         //改变部门
-        mysqli_query($con, "UPDATE `users` SET department = '" . mysqli_escape_string($con, $_REQUEST["depart"]) . "' WHERE id = " . $_SESSION["loginid"]);
-        $_SESSION["message"] = "部门成功更改为" . $_REQUEST["depart"];
+        mysqli_query($con, "UPDATE `users` SET mail = '" . mysqli_escape_string($con, $_REQUEST["email"]) . "' WHERE id = " . $_SESSION["loginid"]);
+        $_SESSION["message"] = "邮箱成功更改为" . $_REQUEST["email"];
         header("Location: manage.php");
     } elseif ($_REQUEST["type"] == "change-password") {
         //改变密码
@@ -107,6 +107,7 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         }
         header("Location: manage.php");
     } elseif ($_REQUEST["type"] == "register-allow") {
+        mysqli_query($con, "DELETE FROM `requests` WHERE type = 'register-allow' AND requestid = " . $_REQUEST["id"]);
         if ($_REQUEST["action"] == "accept") {
             mysqli_query($con, "UPDATE `users` SET actived = 1 WHERE id = " . $_REQUEST["id"]);
             $from_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_SESSION["loginid"]));
@@ -135,20 +136,20 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         header("Location: manage.php");
     } elseif ($_REQUEST["type"] == "upgrade-allow") {
         //升级
+        mysqli_query($con, "DELETE FROM `requests` WHERE type = 'upgrade-to-admin' AND requestid = " . $_REQUEST["id"]);
         if ($_REQUEST["action"] == "accept") {
-            mysqli_query($con, "DELETE FROM `requests` WHERE requestid = " . $_REQUEST["id"]);
             mysqli_query($con, "UPDATE `users` SET accessment = 'admin' WHERE id = " . $_REQUEST["id"]);
             $from_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_SESSION["loginid"]));
             $send_to_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_REQUEST["id"]));
             $format_string = "<h1>您的权限升级申请已经被批准</h1><p>尊敬的" . $send_to_info["realname"] . "，您的账号（ID：" . $send_to_info["id"] . "）的升级申请已经被管理员（ID：" . $from_info['id'] . "）" . $from_info['realname'] . "<b>批准</b>。现在，您可以正式成为中级管理员了。</p>";
             $_SESSION["message"] = "成功批准" . $send_to_info["realname"] . "的请求";
-            send_mail($format_string, $from_info["mail"], "您的账号已被升级");
+            send_mail($format_string, $send_to_info["mail"], "您的账号已被升级");
         } else {
             $from_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_SESSION["loginid"]));
             $send_to_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_REQUEST["id"]));
             $format_string = "<h1>您的权限升级申请已经被批准</h1><p>尊敬的" . $send_to_info["realname"] . "，您的账号（ID：" . $send_to_info["id"] . "）的升级申请被管理员（ID：" . $from_info['id'] . "）" . $from_info['realname'] . "<b>驳回</b>。</p>";
             $_SESSION["message"] = "成功驳回升级权限的请求";
-            send_mail($format_string, $from_info["mail"], "您的账号升级申请被驳回");
+            send_mail($format_string, $send_to_info["mail"], "您的账号升级申请被驳回");
         }
         header("Location: accept.php");
     } elseif ($_REQUEST["type"] == "request-register-allow") {
@@ -160,5 +161,42 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         );
         $_SESSION["message"] = "提交申请成功，请查看邮箱以了解后续进展";
         header("Location: manage.php");
+    } elseif ($_REQUEST["type"] == "queryroom") {
+        //查询邮箱
+        if (mysqli_num_rows(mysqli_query($con, "SELECT * FROM `rooms` WHERE `number`=" . $_REQUEST["roomnumber"])) != 0) {
+            echo "true";
+        } else {
+            echo "false";
+        }
+    } elseif ($_REQUEST["type"] == "change-manage") {
+        //更改管理房间
+        mysqli_query($con, "INSERT INTO `requests` VALUES (NULL," . $_SESSION["loginid"] . ",'change-manage','" . mysqli_escape_string($con, json_encode([
+            "change-to-rooms" => explode(",", $_REQUEST["room-data"])
+        ])) . "',CURRENT_TIMESTAMP())");
+        $_SESSION["message"] = "成功生成更改管理区域的请求，请等待系统管理员批准";
+        header("Location: manage.php");
+    } elseif ($_REQUEST["type"] == "change-manage-allow") {
+        if ($_REQUEST["action"] == "accept") {
+            $new_change = [];
+            foreach (json_decode(mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `requests` where `type` = 'change-manage' AND requestid = " . $_REQUEST["id"]))["param"], true)["change-to-rooms"] as $value) {
+                array_push($new_change, mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `rooms` where `number` = " . $value))["id"]);
+            }
+            mysqli_query($con, "UPDATE `users` SET managepartid = " .
+                (array_count_values($new_change) != 0 ? "'" . implode(",", $new_change) . "'" : "NULL") .
+                " WHERE id = " . $_REQUEST["id"]);
+            $from_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_SESSION["loginid"]));
+            $send_to_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_REQUEST["id"]));
+            $format_string = "<h1>您的权限更改申请已经被批准</h1><p>尊敬的" . $send_to_info["realname"] . "，您的账号（ID：" . $send_to_info["id"] . "）已经被管理员（ID：" . $from_info['id'] . "）" . $from_info['realname'] . "<b>批准</b>。</p>";
+            send_mail($format_string, $from_info["mail"], "您的账号已被批准");
+            $_SESSION["message"] = "成功批准" . $send_to_info["realname"] . "的请求";
+        } else {
+            $from_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_SESSION["loginid"]));
+            $send_to_info = mysqli_fetch_assoc(mysqli_query($con, "SELECT * FROM `users` where id=" . $_REQUEST["id"]));
+            $format_string = "<h1>您的权限更改申请被驳回</h1><p>尊敬的" . $send_to_info["realname"] . "，您的账号（ID：" . $send_to_info["id"] . "）被管理员（ID：" . $from_info['id'] . "）" . $from_info['realname'] . "<b>驳回</b>。如果您需要再次发送申请邮件，请在管理界面申请重新发送。</p>";
+            send_mail($format_string, $from_info["mail"], "您的权限更改申请被驳回");
+            $_SESSION["message"] = "成功驳回" . $send_to_info["realname"] . "的请求";
+        }
+        mysqli_query($con, "DELETE FROM `requests` WHERE type = 'change-manage' AND requestid = " . $_REQUEST["id"]);
+        header("Location: accept.php");
     }
 }
